+++ gui/vad.py
